simplify api, part 1
