+++ pydivert/consts.py
@@ -1,6 +1,4 @@
-import unittest
-
-from pydivert.models import format_structure
+# -*- coding: utf-8 -*-
 # Copyright (C) 2016  Fabio Falcinelli
 #
 # This program is free software: you can redistribute it and/or modify
@@ -15,19 +13,39 @@ from pydivert.models import format_structure
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+from enum import Enum
 
 __author__ = 'fabio'
 
 
-class ModelsTestCase(unittest.TestCase):
-    def setUp(self):
-        pass
+# Divert layers.
+class Layer(Enum):
+    NETWORK = 0
+    NETWORK_FORWARD = 1
 
-    def tearDown(self):
-        pass
 
-    def test_format_structure_raise_exc(self):
-        """
-        Tests the format_structure method
-        """
-        self.assertRaises(ValueError, format_structure, "some_obj")
+# Divert Flag.
+class Flag(Enum):
+    SNIFF = 1
+    DROP = 2
+
+
+# Divert parameters.
+class Param(Enum):
+    QUEUE_LEN = 0  # Packet queue length 1 <default 512 < 8192
+    QUEUE_TIME = 1  # Packet queue time 128 < default 512 < 2048
+
+
+# Direction outbound/inbound
+class Direction(Enum):
+    OUTBOUND = 0
+    INBOUND = 1
+
+
+# Checksums
+class CalcChecksumsOption(Enum):
+    NO_IP_CHECKSUM = 1
+    NO_ICMP_CHECKSUM = 2
+    NO_ICMPV6_CHECKSUM = 4
+    NO_TCP_CHECKSUM = 8
+    NO_UDP_CHECKSUM = 16
